requirements.txt
fastapi[standard]
pydantic-settings
agno
groq
duckduckgo-search
google-genai
sqlmodel
streamlit
requests
python-jose[cryptography]
passlib[bcrypt]